# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "0.0.1",
    "title": "Microservice for Questionnaire Demo Project"
  },
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/start": {
      "get": {
        "x-swagger-router-controller": "start",
        "description": "Start Questionnaire",
        "operationId": "start",
        "tags": [
          "/start"
        ],
        "parameters": [{
            "name": "questionnaireId",
            "in": "header",
            "description": "Id for desired Questionnaire",
            "required": true,
            "type": "string"
          },{
              "name": "sessionId",
              "in": "header",
              "description": "Session Id to allow access to endpoint",
              "required": true,
              "type": "string"
            }],
        "responses": {
          "200": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
            }
          },          
          "500": {
            "description": "Server Error",
            "schema": {
            }
          }
        }
      }
    },
    "/next": {
      "post": {
        "x-swagger-router-controller": "next",
        "description": "Next Questions and Answers in Questionaire",
        "operationId": "next",
        "tags": [
          "/next"
        ],
        "parameters": [{
            "name": "sessionId",
            "in": "header",
            "description": "SessionId for User",
            "required": true,
            "type": "string"
          },
          {
              "name": "questionnaireId",
              "in": "header",
              "description": "Id for desired Questionnaire",
              "required": true,
              "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "answer",
            "required": true,
            "schema": {
              "$ref": "#/definitions/question_v001_schema"
            }
          }],
        "responses": {
          "200": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
            }
          }
        }
      }
    },
    "/summary": {
      "get": {
        "x-swagger-router-controller": "summary",
        "description": "All Questions and Answers from Questionaire",
        "operationId": "summary",
        "tags": [
          "/summary"
        ],
        "parameters": [{
            "name": "sessionId",
            "in": "header",
            "description": "SessionId for User",
            "required": true,
            "type": "string"
        }],
        "responses": {
          "200": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request",
            "schema": {
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
            }
          },
          "500": {
            "description": "Server Error",
            "schema": {
            }
          }
        }
      }
    }
  },
  "definitions": {
    "question_v001_schema": {
      "properties": {
        "questionId": {
          "type": "string",
          "example": "5a4b1b59af7001a735b50c80"
        },
        "answerId": {
          "type": "string",
          "example": "5a4b1b59af7001a735b50c80"
        }
      }
    }
  }
}